+++ library/include/ProceduralPath.h
@@ -160,8 +160,8 @@ public:
 	const Ogre::Vector3& getPoint(int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/** Gets the number of segments in the path
@@ -169,7 +169,7 @@ public:
 	 */
 	int getSegCount() const
 	{
-		return (mPoints.size()-1) + (mClosed?1:0);
+		return (int)(mPoints.size()-1) + (mClosed?1:0);
 	}
 
 	/**
@@ -391,7 +391,7 @@ public:
 
 	unsigned int getPathCount() const
 	{
-		return mPaths.size();
+		return (unsigned int)mPaths.size();
 	}
 
 	Path getPath(unsigned int i) const
